Add toggle to chart to show/hide the chart
This is a little weird, we render the chart's
container as a server component so it can fetch
data, but the chart itself is a client component.

The toggle button needs to be a client component,
so we have to wrap it in a client component and
pass the server component as a child. 🤷‍♂️